Revert "Changing stepTime type to double"
This reverts commit 1f1574aead7bf7b95aae04ea5702941bfb51ef71.

+++ src/ui/create_simulation_dialog.cpp
@@ -222,7 +222,7 @@ void CreateSimulationDialog::accept() {
     simulation->setSimulationType(Simulation::getSimulationTypesMap().key(ui->cbxType->currentText()));
     simulation->setInitialTime(time.toTime_t());
     simulation->setPeriod(ui->edtPeriod->text().toDouble());
-    simulation->setStepTime(ui->edtStepTime->text().toDouble());
+    simulation->setStepTime(ui->edtStepTime->text().toInt());
     simulation->setMinimumVerticalLimit(ui->edtMinLimit->text().toDouble());
     simulation->setMaximumVerticalLimit(ui->edtMaxLimit->text().toDouble());
